move converting functions into a helper class; enhance convertDate function; housekeeping
